% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directed_example_net.R
\docType{data}
\name{directed_example_net}
\alias{directed_example_net}
\title{Directed Example Network}
\format{
An `network` object with 5 nodes and 8 edges.
}
\usage{
directed_example_net
}
\description{
A one-mode directed network of 5 individuals represented as a `network` object.
}
\details{
This is a fictitious network used as an example.


Vertex attributes:
- `vertex.names`: Name of individual
}
\examples{
data( directed_example_net )
sna::gplot( directed_example_net )
# coerce to a matrix
as.matrix( directed_example_net )
}
\keyword{datasets}
