% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rescaling.R
\name{edge.rescale}
\alias{edge.rescale}
\title{Rescale Numeric Matrix for Plotting Edge Width}
\usage{
edge.rescale(uni_mat, low, high)
}
\arguments{
\item{uni_mat}{A matrix of weights to use for edge widths.}

\item{low}{Minimum of target range.}

\item{high}{Maximum of target range.}
}
\value{
Rescaled numeric matrix.
}
\description{
Rescale Numeric Matrix for Plotting Edge Width
}
