% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rescaling.R
\name{edge.shade}
\alias{edge.shade}
\title{Rescale Numeric Matrix for an Edge Attribute to a New Range for Plotting}
\usage{
edge.shade(uni_mat)
}
\arguments{
\item{uni_mat}{A matrix of weights to use for edge shading, where higher
values represent darker edges.}
}
\value{
Rescaled numeric matrix.
}
\description{
Rescale Numeric Matrix for an Edge Attribute to a New Range for Plotting
}
