% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_egocentric.R
\name{egosd_r}
\alias{egosd_r}
\title{Egocentric SD (Receive Network)}
\usage{
egosd_r(adj, var)
}
\arguments{
\item{adj}{A square adjacency matrix.}

\item{var}{A numeric vector of attribute values for each node.}
}
\value{
A numeric vector of egocentric standard deviations for each ego.
}
\description{
Calculates the standard deviation of a variable among alters who nominate each ego.
}
