% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroin_dealing_net.R
\docType{data}
\name{heroin_dealing_net}
\alias{heroin_dealing_net}
\title{Heroin Dealing Network}
\format{
An `network` object with 38 nodes and 1168 edges.
}
\source{
Data retrieved from \url{https://sites.google.com/site/ucinetsoftware/datasets/covert-networks/cocaine-dealing-natarajan}
}
\usage{
heroin_dealing_net
}
\description{
A one-mode directed network of 38 individuals represented as a `network` object.
}
\details{
The network comes from an investigation into to a large heroin trafficking organization in New York City.
The ties are communications through police wiretaps and represented as directed binary relations.


Vertex attributes:
- `vertex.names`: Name of individual
}
\examples{
sna::gplot( heroin_dealing_net )
}
\references{
Natarajan, M. (2006). Understanding the structure of a large heroin distribution network:
A quantitative analysis of qualitative data. *Journal of Quantitative Criminology*, 22(2), 171–192.
\url{https://doi.org/10.1007/s10940-006-9007-x}
}
\keyword{datasets}
