% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafia_calls_net.R
\docType{data}
\name{mafia_calls_net}
\alias{mafia_calls_net}
\title{Mafia Calls Network}
\format{
An `network` object with 98 nodes and 123 edges.
}
\source{
Data retrieved from \url{https://github.com/lcucav/criminal-nets/blob/master/dataset/Montagna_phonecalls_edgelist.csv}
}
\usage{
mafia_calls_net
}
\description{
A one-mode, undirected, binary network of individuals and phone calls represented as a `network` object.
}
\details{
The network represents phone calls among members of the Sicilian Mafia in the first decade of
the 2000s.


Vertex attributes:
- `vertex.names`: Names
}
\examples{
data( mafia_calls_net )
sna::gplot( mafia_calls_net, gmode = "graph" )
}
\references{
Cavallaro, L., Ficara, A., De Meo, P., Fiumara, G., Catanese, S., et al. (2020). Disrupting resilient criminal networks through data analysis: The case of Sicilian Mafia. *PLOS ONE*, 15(8), e0236476.
\url{https://doi.org/10.1371/journal.pone.0236476}
}
\keyword{datasets}
