% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_events_net.R
\docType{data}
\name{officer_event_net}
\alias{officer_event_net}
\title{Officer Event Network}
\format{
A `network` object with 234 nodes (81 individuals and 153 events) and 346 edges.
}
\usage{
officer_event_net
}
\description{
A two-mode network of police officers and incidents represented as a `network` object.
}
\details{
The network comes from an study of officers co-participation in incidents.


Vertex attributes:
- `vertex.names`: Name of individual
}
\examples{
data( officer_event_net )
}
\references{
Young, J. T. N., & Ready, J. T. (2015). Diffusion of ideas and technology: The role of networks in influencing the endorsement and use of on-officer video cameras. *Journal of Contemporary Criminal Justice*, 31(3), 243–261.
\url{https://doi.org/10.1177/1043986214553380}
}
\keyword{datasets}
