% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paul_revere_net.R
\docType{data}
\name{paul_revere_net}
\alias{paul_revere_net}
\title{Paul Revere Conspiracy Network}
\format{
A `network` object with 261 nodes (254 individuals and 7 events) and 334 edges.
}
\usage{
paul_revere_net
}
\description{
A two-mode network of individuals and locations represented as a `network` object.
}
\details{
The Paul Revere conspiracy dataset concerns relationships between people and their
affiliations with different organizations in Boston. The dataset refers to Paul Revere,
who was responsible for organizing a local militia of Boston's revolutionary
movement (see http://en.wikipedia.org/wiki/Sons_of_Liberty ). The dataset was analysed
by Kieran Healy of Duke University. This dataset has been reconstructed by looking
at the information presented in the appendix of the book "Paul Revere's Ride"
published by David Fischer (1994).


Vertex attributes:
- `names`: Name of individuals and locations
}
\examples{
data( paul_revere_net )
}
\references{
Fischer, D. (1994). "Paul Revere's Ride", Oxford University Press.
\url{http://www.casos.cs.cmu.edu/tools/datasets/external/paulrevere/}
}
\keyword{datasets}
