% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{roster_to_adjacency}
\alias{roster_to_adjacency}
\title{Roster Data to Adjacency Matrix}
\usage{
roster_to_adjacency(input_dat)
}
\arguments{
\item{input_dat}{an object where the first column is a set of ids
and the subsequent columns are roster nominations.}
}
\value{
An adjacency matrix of class \code{matrix}.
}
\description{
Takes data in roster format and creates an adjacency matrix.
}
