% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undirected_example_net.R
\docType{data}
\name{undirected_example_net}
\alias{undirected_example_net}
\title{Undirected Example Network}
\format{
An `network` object with 5 nodes and 5 edges.
}
\usage{
undirected_example_net
}
\description{
A one-mode undirected network of 5 individuals represented as a `network` object.
}
\details{
This is a fictitious network used as an example.


Vertex attributes:
- `vertex.names`: Name of individual
}
\examples{
data( undirected_example_net )
sna::gplot( undirected_example_net, gmode = "graph" )
# coerce to a matrix
as.matrix( undirected_example_net )
}
\keyword{datasets}
