% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rescaling.R
\name{vertex_rescale}
\alias{vertex_rescale}
\title{Rescale Numeric Values for a Vertex Attribute to a New Range for Plotting}
\usage{
vertex_rescale(x, low, high)
}
\arguments{
\item{x}{A numeric vector.}

\item{low}{Minimum of target range.}

\item{high}{Maximum of target range.}
}
\value{
Rescaled numeric vector.
}
\description{
Rescale Numeric Values for a Vertex Attribute to a New Range for Plotting
}
