% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocaine_dealing_net.R
\docType{data}
\name{cocaine_dealing_net}
\alias{cocaine_dealing_net}
\title{Cocaine Dealing Network}
\format{
An `network` object with 28 nodes and 40 edges.
}
\source{
Data retrieved from \url{http://www.casos.cs.cmu.edu/tools/datasets/external/index.php}
}
\usage{
cocaine_dealing_net
}
\description{
A one-mode network of 28 individuals represented as a `network` object.
}
\details{
The network comes from an investigation into to a large cocaine trafficking organization in New York City.
The ties are communications through police wiretaps and represented as directed binary relations.


Vertex attributes:
- `vertex.names`: Name of individual
}
\examples{
sna::gplot( cocaine_dealing_net )
}
\references{
Natarajan, M. (2000). *Understanding the Structure of a Drug Trafficking
Organization: A Conversational Analysis*. In M. Natarajan & M. Hough (Eds.),
*Illegal Drug Markets: From Research to Prevention Policy* (Crime Prevention
Studies, Vol. 11, pp. 273--298). Criminal Justice Press.
}
\keyword{datasets}
