% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phx_arrest_2024_net.R
\docType{data}
\name{phx_arrest_2024_net}
\alias{phx_arrest_2024_net}
\title{Phoenix 2024 Arrest Network}
\format{
A `network` object with 3,052 nodes (2,029 individuals and 1,023 events) and 3,314 edges.
}
\usage{
phx_arrest_2024_net
}
\description{
A mode network of individuals and events represented as a `network` object.
}
\details{
Vertex attributes:
- `vertex.names`: Name of individuals and locations
}
\examples{
data( phx_arrest_2024_net )
}
\references{
Phoenix Open Data Portal
\url{https://www.phoenixopendata.com/dataset/arrests}
}
\keyword{datasets}
