% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rescaling.R
\name{vertex_rescale}
\alias{vertex_rescale}
\title{Rescale Numeric Values for a Vertex Attribute to a New Range for Plotting}
\usage{
vertex_rescale(x, low, high)
}
\arguments{
\item{x}{A numeric vector.}

\item{low}{Minimum of target range.}

\item{high}{Maximum of target range.}
}
\value{
Rescaled numeric vector.
}
\description{
This function takes a vector of values that need to be rescaled for a plot.
If the vector is a constant, it will return an error indicating that there
is not a need to rescale the vector.
}
